@@ -9,12 +9,12 @@
     "express-session": "^1.17.2",
     "mongodb": "^4.4.0",
     "radium": "^0.26.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-animations": "^1.0.0",
     "react-bootstrap": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-ionicons": "^4.2.0",
-    "react-redux": "^7.2.4",
+    "react-redux": "^8.0.5",
     "react-router-dom": "^6.2.1",
     "react-s-alert": "^1.4.1",
     "react-scripts": "^5.0.0",
@@ -22,7 +22,7 @@
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^2.0.3",
     "sha256": "^0.2.0",
-    "uuid": "^8.3.2"
+    "uuid": "^9.0.0"
   },
   "description": "This is an example of signin and signuot on reactjs and mongodb",
   "scripts": {
