@@ -1,9 +1,10 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+// import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 // import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import App from './components/App';
-import registerServiceWorker from './registerServiceWorker';
+// import registerServiceWorker from './registerServiceWorker';
 
 // react-router-dom
 import { BrowserRouter } from 'react-router-dom'
@@ -17,11 +18,11 @@ import { Provider } from 'react-redux'
 import 'semantic-ui-css/semantic.min.css';
 
 const logger = store => next => action => {
-  console.group(action.type)
-  console.info('dispatching', action)
+  // console.group(action.type)
+  // console.info('dispatching', action)
   let result = next(action)
-  console.log('next state', store.getState())
-  console.groupEnd(action.type)
+  // console.log('next state', store.getState())
+  // console.groupEnd(action.type)
   return result
 }
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
@@ -32,12 +33,23 @@ const store = createStore(
   )
 )
 
-ReactDOM.render(
+// ReactDOM.render(
+//   <Provider store={store}>
+//     <BrowserRouter>
+//       <App />
+//     </BrowserRouter>
+//   </Provider>,
+//   document.getElementById('root')
+// );
+
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
   <Provider store={store}>
     <BrowserRouter>
       <App />
     </BrowserRouter>
-  </Provider>,
-  document.getElementById('root')
+  </Provider>
 );
-registerServiceWorker();
+
+// registerServiceWorker();
